# Optimize `isnan` (2-5x faster)