--- a/datafusion/functions/src/math/nans.rs
+++ b/datafusion/functions/src/math/nans.rs
@@ -17,11 +17,11 @@
 
 //! Math function: `isnan()`.
 
-use arrow::datatypes::DataType;
-use datafusion_common::{exec_err, DataFusionError, Result};
+use arrow::datatypes::{DataType, Float32Type, Float64Type};
+use datafusion_common::{exec_err, Result};
 use datafusion_expr::{ColumnarValue, TypeSignature};
 
-use arrow::array::{ArrayRef, BooleanArray, Float32Array, Float64Array};
+use arrow::array::{ArrayRef, AsArray, BooleanArray};
 use datafusion_expr::{ScalarUDFImpl, Signature, Volatility};
 use std::any::Any;
 use std::sync::Arc;
@@ -72,20 +72,15 @@ impl ScalarUDFImpl for IsNanFunc {
         let args = ColumnarValue::values_to_arrays(args)?;
 
         let arr: ArrayRef = match args[0].data_type() {
-            DataType::Float64 => Arc::new(make_function_scalar_inputs_return_type!(
-                &args[0],
-                self.name(),
-                Float64Array,
-                BooleanArray,
-                { f64::is_nan }
-            )),
-            DataType::Float32 => Arc::new(make_function_scalar_inputs_return_type!(
-                &args[0],
-                self.name(),
-                Float32Array,
-                BooleanArray,
-                { f32::is_nan }
-            )),
+            DataType::Float64 => Arc::new(BooleanArray::from_unary(
+                args[0].as_primitive::<Float64Type>(),
+                f64::is_nan,
+            )) as ArrayRef,
+
+            DataType::Float32 => Arc::new(BooleanArray::from_unary(
+                args[0].as_primitive::<Float32Type>(),
+                f32::is_nan,
+            )) as ArrayRef,
             other => {
                 return exec_err!(
                     "Unsupported data type {other:?} for function {}",