## Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and enhancements and this helps us generate change logs for our releases. You can link an issue to this PR using the GitHub syntax. For example `Closes #123` indicates that this PR will close issue #123.
-->

Closes #.

## Rationale for this change

<!--
 Why are you proposing this change? If this is already explained clearly in the issue then this section is not needed.
 Explaining clearly why changes are proposed helps reviewers understand your changes and offer better suggestions for fixes.  
-->

Apply same idea from https://github.com/apache/datafusion/pull/12881 to `isnan`.

## What changes are included in this PR?

<!--
There is no need to duplicate the description in the issue here but it is sometimes worth providing a summary of the individual changes in this PR.
-->

- Add bench for is_nan
- Optimize implementation

## Are these changes tested?

<!--
We typically require tests for all PRs in order to:
1. Prevent the code from being accidentally broken by subsequent changes
2. Serve as another way to document the expected behavior of the code

If tests are not included in your PR, please explain why (for example, are they covered by existing tests)?
-->

Yes, with existing tests.

## Are there any user-facing changes?

Yes, faster `is_nan` function.

```
isnan f32 array: 1024   time:   [794.46 ns 795.95 ns 797.51 ns]
                        change: [-60.752% -60.453% -60.230%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild

isnan f64 array: 1024   time:   [694.70 ns 698.25 ns 702.33 ns]
                        change: [-66.246% -66.029% -65.767%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 10 outliers among 100 measurements (10.00%)
  6 (6.00%) high mild
  4 (4.00%) high severe

isnan f32 array: 4096   time:   [2.5207 µs 2.5282 µs 2.5358 µs]
                        change: [-64.245% -64.126% -64.015%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high mild

isnan f64 array: 4096   time:   [2.0779 µs 2.0812 µs 2.0847 µs]
                        change: [-74.140% -74.042% -73.950%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  3 (3.00%) high mild
  2 (2.00%) high severe

isnan f32 array: 8192   time:   [4.9547 µs 4.9730 µs 4.9971 µs]
                        change: [-74.219% -74.059% -73.933%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) high mild
  2 (2.00%) high severe

isnan f64 array: 8192   time:   [4.0478 µs 4.0562 µs 4.0652 µs]
                        change: [-82.993% -82.947% -82.900%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  3 (3.00%) high mild
  1 (1.00%) high severe
```

<!--
If there are user-facing changes then we may require documentation to be updated before approving the PR.
-->

<!--
If there are any breaking changes to public APIs, please add the `api change` label.
-->